+++ frontend/src/pages/WorkflowitemDetails/actions.js
@@ -1,14 +1,20 @@
-export const FETCH_WORKFLOWITEM_HISTORY = "FETCH_WORKFLOWITEM_HISTORY";
-export const FETCH_WORKFLOWITEM_HISTORY_SUCCESS = "FETCH_WORKFLOWITEM_HISTORY_SUCCESS";
+export const SET_TOTAL_WORKFLOWITEM_HISTORY_ITEM_COUNT = "SET_TOTAL_WORKFLOWITEM_HISTORY_ITEM_COUNT";
+export const FETCH_NEXT_WORKFLOWITEM_HISTORY_PAGE = "FETCH_NEXT_WORKFLOWITEM_HISTORY_PAGE";
+export const FETCH_NEXT_WORKFLOWITEM_HISTORY_PAGE_SUCCESS = "FETCH_NEXT_WORKFLOWITEM_HISTORY_PAGE_SUCCESS";
 
-export function fetchWorkflowitemHistory(projectId, subprojectId, workflowitemId, offset, limit, showLoading = false) {
+export function setTotalHistoryItemCount(count) {
   return {
-    type: FETCH_WORKFLOWITEM_HISTORY,
+    type: SET_TOTAL_WORKFLOWITEM_HISTORY_ITEM_COUNT,
+    count
+  };
+}
+
+export function fetchNextWorkflowitemHistoryPage(projectId, subprojectId, workflowitemId, showLoading = false) {
+  return {
+    type: FETCH_NEXT_WORKFLOWITEM_HISTORY_PAGE,
     projectId,
     subprojectId,
     workflowitemId,
-    offset,
-    limit,
     showLoading
   };
 }